feat(download): add download queue optimization (Phase 22)
- Add queue_position field to DownloadTask for explicit queue ordering
- Update try_start_next_queued() to use queue_position as tiebreaker
  (lower position = higher priority within same DownloadPriority level)
- Add queue manipulation methods: set/get_queue_position,
  move_task_up/down, move_task_to_top/bottom
- Persist queue_position in task_queue.json (serde default for backward compat)
- Add /dlqmove CLI command (up/down/top/bottom)
- Fix pre-existing clippy warnings in disk_monitor.rs
- 215 tests passed, zero warnings

+++ ipmsg-torrent/crates/download/src/disk_monitor.rs
@@ -60,7 +60,7 @@ pub fn get_available_space(path: &Path) -> Result<u64, DiskSpaceError> {
         let stat = unsafe { stat.assume_init() };
         // f_bavail is available blocks for non-privileged users
         // f_frsize is the fundamental filesystem block size
-        Ok(stat.f_bavail as u64 * stat.f_frsize as u64)
+        Ok(stat.f_bavail * stat.f_frsize)
     }
 
     #[cfg(not(unix))]
